move useStreamDeckSelection into main async module
