+++ src/app/stream-deck/OpenOnStreamDeckButton/OpenOnStreamDeckButton.tsx
@@ -4,7 +4,7 @@ import ActionButton from 'app/item-actions/ActionButton';
 import { BucketHashes } from 'data/d2/generated-enums';
 import streamDeckIcon from 'images/streamDeck.svg';
 import { useMemo } from 'react';
-import { useStreamDeckSelection } from '../useStreamDeckSelection';
+import { useStreamDeckSelection } from '../stream-deck';
 import styles from './OpenOnStreamDeckButton.m.scss';
 
 export default function OpenOnStreamDeckButton({ item, label }: { item: DimItem; label: boolean }) {
